requirements.txt
# LLM APIs
anthropic>=0.8.0

# Agent Framework
pydantic-ai==0.0.18

# Environment Variables
python-dotenv>=1.0.0
